refactor: login and registration page
according to scss modules
