+++ frontend/src/css/main.scss
@@ -12,6 +12,5 @@
 // PAGES
 //@import "./pages/code-of-conduct";
 //@import "./pages/about-us";
-@import "./pages/registration";
 @import "./pages/project-listing";
 @import "./pages/admin-profile";
