+++ frontend/src/pages/Registration/registration.module.scss
@@ -1,3 +1,5 @@
+@import "../../css/settings/_global";
+
 .registration-page{
     background-color:$yellow-bg;
